Update Firefox_56.0.2_Linux_64_MX-install.sh
# Firefox--Custom-Installation-Debian9-Slackware14.2--
This is a Bash script in which Mozilla Firefox is custom installed in Debian 9 and Slackware 14.2, with more complete control of the installation site. Code optimization will be implemented gradually. The code can be downloaded, used or modified.

- [Update date: 26/072017] firefox-56.0.2
- [Update date: 24/10/2017] firefox-56.0.1

<b>Installation</b>

To install only you have to give execute permissions as root and execute:

<code>#chmod +x file<code>

<code>#./file<code>
  
  <b>Unistall</b>
  
  Run the file "remove.sh" as the root user:
  #bash /installation directory/firefox/remove.sh